% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pseudolikelihood_functions.R
\name{compute_log_pseudoposterior}
\alias{compute_log_pseudoposterior}
\title{Compute the proportional log pseudoposterior density value.}
\usage{
compute_log_pseudoposterior(x, sigma, prior_var = 1)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{prior_var}{The variance of the prior distribution stipulated on the
Ising model parameters. Currently a normal distribution is used for all
model parameters, with a mean equal to zero and a variance equal to
\code{prior_var}. Defaults to \code{1}.}
}
\value{
A single numeric value.
}
\description{
The function \code{\link{compute_log_pseudoposterior}} computes the proportional
log pseudoposterior density if \code{prior_var} is finite, and computes the
log pseudolikelihood if \code{prior_var} is set to \code{Inf}.
}
\details{
The pseudolikelihood is the product of all full-conditional distributions of
the full Ising model. Used by \code{\link{optimize_pseudoposterior}} to check for
convergence of the Newton Raphson procedure.
}
