% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emvs_functions.R
\name{compute_log_pseudoposterior_emvs}
\alias{compute_log_pseudoposterior_emvs}
\title{Compute the proportional log joint pseudoposterior density value.}
\usage{
compute_log_pseudoposterior_emvs(
  x,
  sigma,
  theta = 0.5,
  prior_var_intercepts = 1,
  spike_var,
  slab_var,
  hierarchical = FALSE,
  alpha = 1,
  beta = 1
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{theta}{The prior inclusion probability. The value \code{theta = 0.5},
in combination with \code{hierarchical = FALSE} stipulates a uniform prior
on the space of network structures.}

\item{prior_var_intercepts}{The variance of the prior distribution
stipulated on the Ising model's threshold parameters. Currently a normal
distribution is used for all threshold parameters, with a mean equal to
zero and a variance equal to \code{prior_var_intercepts}. Defaults to
\code{1}.}

\item{spike_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}

\item{slab_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}

\item{hierarchical}{Logical. If TRUE, a beta prior distribution is
imposed on the prior inclusion probability \code{theta} with
hyperparameters \code{alpha} and \code{beta}. A uniform prior on the
inclusion probability, a beta with \code{alpha = beta = 1}, stipulates a
uniform prior on network structure complexity.}

\item{alpha, beta}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}
}
\value{
A single numeric value.
}
\description{
The function \code{\link{compute_log_pseudoposterior_emvs}} computes the
proportional log joint pseudoposterior density.
}
\details{
Used by \code{\link{edge_screening_emvs}} to check for convergence of the
Newton Raphson procedure.
}
