% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pseudolikelihood_functions.R
\name{compute_standard_deviation}
\alias{compute_standard_deviation}
\title{Compute the asymptotic pseudoposterior standard deviations.}
\usage{
compute_standard_deviation(x, sigma, prior_var = 1)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{prior_var}{The variance of the prior distribution stipulated on the
Ising model parameters. Currently a normal distribution is used for all
model parameters, with a mean equal to zero and a variance equal to
\code{prior_var}. Defaults to \code{1}.}
}
\value{
A \code{p} by \code{p} numeric matrix with standard deviations for
  the pairwise association parameters on the off-diagonal elements and for
  the threshold paramaters on the diagonal elements.
}
\description{
The function \code{compute_standard_deviation} computes the pseudoposterior
standard deviations based on an asymptotic normal approximation. If
\code{prior_var} is set to \code{Inf}, it returns the standard errors of the
pseudolikelihood parameters instead.
}
\details{
The function \code{compute_standard_deviation} calls
\code{\link{invert_hessian}} to produce the standard deviations
(standard errors).
}
