% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emvs_functions.R
\name{edge_screening_emvs}
\alias{edge_screening_emvs}
\title{EM edge inclusion procedure.}
\usage{
edge_screening_emvs(
  x,
  sigma,
  theta = 0.5,
  prior_var_intercepts = 1,
  spike_var,
  slab_var,
  hierarchical = FALSE,
  alpha = 1,
  beta = 1,
  iteration_max = 100
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements. Optional. Can be used to specify starting values.}

\item{theta}{The prior inclusion probability. The value \code{theta = 0.5},
in combination with \code{hierarchical = FALSE} stipulates a uniform prior
on the space of network structures.}

\item{prior_var_intercepts}{The variance of the prior distribution
stipulated on the Ising model's threshold parameters. Currently a normal
distribution is used for all threshold parameters, with a mean equal to
zero and a variance equal to \code{prior_var_intercepts}. Defaults to
\code{1}.}

\item{spike_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}

\item{slab_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}

\item{hierarchical}{Logical. If TRUE, a beta prior distribution is
imposed on the prior inclusion probability \code{theta} with
hyperparameters \code{alpha} and \code{beta}. A uniform prior on the
inclusion probability, a beta with \code{alpha = beta = 1}, stipulates a
uniform prior on network structure complexity.}

\item{alpha}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}

\item{beta}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}

\item{iteration_max}{The maximum number of Newton-Raphson iterations used in
optimization. Defaults to \code{1e2}. A warning is issued if procedure has
not converged in \code{iteration_max} iterations.}
}
\value{
A list containing the \code{p} by \code{p} matrices \code{sigma} and
  \code{gamma}. The matrix \code{sigma} is a numeric matrix with pairwise
  association estimates on the off-diagonal elements and threshold
  estimates on the diagonal elements. The matrix \code{gamma} contains the
  expected values of edge inclusion variables (i.e., the local posterior
  probability of edge inclusion. If \code{hierarchical = TRUE}, the modal
  estimate of the prior inclusion probability \code{theta} is also provided.
}
\description{
The function \code{edge_screening_emvs} estimates the
(local) posterior probability that an edge should be included using the EM
algorithm.
}
\details{
The function \code{\link{screen_edges}} is a wrapper for
\code{edge_screening_emvs}.
}
