% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{indexing}
\alias{indexing}
\title{Indexing matrix for converting parameter matrix to vector.}
\usage{
indexing(p)
}
\arguments{
\item{p}{The number of rows and columns of the parameter matrix \code{sigma}.}
}
\value{
A \code{p * (p + 1) / 2} by \code{3} matrix of integers. The first
  column contains the index of the vector \code{eta}, the second and third
  column the corresponding row and column in \code{sigma}.
}
\description{
The function \code{indexing} is used to convert the \code{p} by \code{p}
parameter matrix \code{sigma} to a \code{p * (p + 1) / 2} vector \code{eta}
during optimization.
}
\details{
The function \code{indexing} is used by
\code{\link{compute_standard_deviation}},
\code{\link{optimize_pseudoposterior}},
and ....
}
