% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pseudolikelihood_functions.R
\name{invert_hessian}
\alias{invert_hessian}
\title{Invert the log pseudoposterior Hessian matrix.}
\usage{
invert_hessian(x, sigma, index, prior_var = 1)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{index}{An indexing matrix used to convert the matrix \code{sigma} to a
vector \code{eta} during optimization. See \code{\link{indexing}}.}

\item{prior_var}{The variance of the prior distribution stipulated on the
Ising model parameters. Currently a normal distribution is used for all
model parameters, with a mean equal to zero and a variance equal to
\code{prior_var}. Defaults to \code{1}.}
}
\value{
A \code{p(p-1)/2} by \code{p(p-1)/2} numeric matrix.
}
\description{
The function \code{invert_hessian} computes the inverse Hessian of the log
pseudoposterior distribution.
}
\details{
The function \code{invert_hessian} is used by
\code{multidimensional_update} for the multidimensional Newton Raphson step,
and by \code{\link{compute_standard_deviation}} to compute the
pseudoposterior standard deviations.
}
