% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emvs_functions.R
\name{multidimensional_update_emvs}
\alias{multidimensional_update_emvs}
\title{Multidimensional Newton Raphson joint pseudoposterior updates for EMVS.}
\usage{
multidimensional_update_emvs(
  x,
  sigma,
  gamma,
  index,
  suff_stat,
  prior_var_intercepts = 1,
  spike_var,
  slab_var
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{gamma}{A \code{p} by \code{p} numeric matrix with expected inclusion
probabilities for the pairwise association parameters (i.e., edges).}

\item{index}{An indexing matrix used to convert the matrix \code{sigma} to a
vector \code{eta} during optimization. See \code{\link{indexing}}.}

\item{suff_stat}{A \code{p} by \code{p} matrix of sufficient statistics.}

\item{prior_var_intercepts}{The variance of the prior distribution
stipulated on the Ising model's threshold parameters. Currently a normal
distribution is used for all threshold parameters, with a mean equal to
zero and a variance equal to \code{prior_var_intercepts}. Defaults to
\code{1}.}

\item{spike_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}

\item{slab_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters.}
}
\value{
A \code{p} by \code{p} numeric matrix with updates pairwise
  association estimates on the off-diagonal elements and updated threshold
  estimates on the diagonal elements.
}
\description{
The function \code{multidimensional_update_emvs} updates all pseudoposterior
values for the Ising model parameters at once.
}
\details{
The function \code{multidimensional_update_emvs} is used by
\code{edge_screening_emvs} to optimize the Ising model parameters.
}
