% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pseudolikelihood_functions.R
\name{onedimensional_update}
\alias{onedimensional_update}
\title{One dimensional Newton Raphson pseudoposterior updates.}
\usage{
onedimensional_update(x, sigma, suff_stat, prior_var = 1)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements.}

\item{suff_stat}{A \code{p} by \code{p} matrix of sufficient statistics.}

\item{prior_var}{The variance of the prior distribution stipulated on the
Ising model parameters. Currently a normal distribution is used for all
model parameters, with a mean equal to zero and a variance equal to
\code{prior_var}. Defaults to \code{1}.}
}
\value{
A \code{p} by \code{p} numeric matrix with updates pairwise
  association estimates on the off-diagonal elements and updated threshold
  estimates on the diagonal elements.
}
\description{
The function \code{onedimensional_update} updates all pseudoposterior values
one at a time.
}
\details{
The function \code{onedimensional_update} is used by
\code{\link{optimize_pseudoposterior}} to provide reasonable starting values. After
which it runs \code{\link{multidimensional_update}}.
}
