% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_pseudolikelihood_functions.R
\name{optimize_pseudoposterior}
\alias{optimize_pseudoposterior}
\title{Optimization of the Ising model's pseudoposterior distribution}
\usage{
optimize_pseudoposterior(x, prior_var = 1, iteration_max = 100)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{prior_var}{The variance of the prior distribution stipulated on the
Ising model parameters. Currently a normal distribution is used for all
model parameters, with a mean equal to zero and a variance equal to
\code{prior_var}. Defaults to \code{1}.}

\item{iteration_max}{The maximum number of Newton-Raphson iterations used in
optimization. Defaults to \code{1e2}. A warning is issued if procedure has
not converged in \code{iteration_max} iterations.}
}
\value{
A \code{p} by \code{p} numeric matrix with pairwise association
  estimates on the off-diagonal elements and threshold estimates on the
  diagonal elements.
}
\description{
The function \code{optimize_pseudoposterior} can be used to optimize the
Ising model's pseudoposterior distribution. It provides the posterior modes
and of both the thresholds and pairwise associations. The function produces
maximum pseudolikelihood estimates whenever \code{prior_var} is set to
\code{Inf}.
}
\details{
Optimization is performed with respect to all \code{p} full-conditional
distributions at once.
}
\seealso{
\code{\link{fit_pseudoposterior}}
}
