% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_edge_BF}
\alias{plot_edge_BF}
\title{Inclusion evidence of edges plot.}
\usage{
plot_edge_BF(output, evidence_thresh = 10, ...)
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{evidence_thresh}{numeric larger 1. Specifies the inclusion Bayes Factor above which
edges are considered included. Default set to 10. The higher, the more conservative the evidence plot.}

\item{...}{Additional arguments to be passed on to qgraph}
}
\value{
Function returns a network with edges indicating evidential strength for in-/exclusion.
}
\description{
The function creates a network showing the evidential strength for inclusion /exclusion. 
Red edges indicate evidence for exclusion, blue edges evidence for inclusion, and grey edges 
absence of evidence with a Bayes Factor below the evidence threshold.
}
\details{
This function uses \code{qgraph} for the visualization of the network. 
Additional arguments from the \code{qgraph} package can be passed on within this function.
}
