% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_mpm}
\alias{plot_mpm}
\title{Median probability model plot.}
\usage{
plot_mpm(output, exc_prob = 0.5, ...)
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{exc_prob}{Number between 0 and 1. Specifies minimum inclusion probability for edges
to be depicted in the network.}

\item{...}{Additional arguments to be passed on to qgraph}
}
\description{
The function depicts the median probability structure of the Ising model, eliminating edges 
with an inclusion probability below the specified \code{exc_prob}-parameter 
(i.e., default set to exc_prob 0.5). Within the network, nodes represent the 
variables of the dataset and lines the respective node interaction. The thicker 
the line the stronger the interaction between two nodes.
}
\details{
This function uses \code{qgraph} for the visualization of the network. 
Additional arguments from the \code{qgraph} package can be passed on within this function.
}
