% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_parameter_distribution}
\alias{plot_parameter_distribution}
\title{Plot posterior distribution of the parameter samples.}
\usage{
plot_parameter_distribution(output, parameter = c("all"))
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{parameter}{A vector stating which parameters of the model should be plotted. 
Options "all", "sigma", and "mu". Default is set to "all"}
}
\value{
Depicts density plots of parameters' posterior distribution.
}
\description{
The function plots the density of the posterior distribution of the parameter samples.
}
