% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_parameter_HDI}
\alias{plot_parameter_HDI}
\title{Highest density interval of parameters plot.}
\usage{
plot_parameter_HDI(output, thresholds = F)
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{thresholds}{binary Determines whether the threshold parameters should be plotted}
}
\value{
Forest plot depicting the median of parameters and their respective 95% HDI
}
\description{
The function plots the 95% highest density interval (HDI) of the posterior sample of the Ising
parameters. Dots represent the median of the posterior distribution and lines the respective 
95% HDI.
}
\details{
The function is based in the ggplot2 environment. Additional arguments can be passed on 
using the common "+" notation and a respective ggplot command.
}
