% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_structure}
\alias{plot_structure}
\title{Most probable structures plot.}
\usage{
plot_structure(output, top_n = 3, ...)
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{top_n}{Number of most probable structures plotted.}

\item{...}{Additional arguments to be passed on to qgraph}
}
\value{
Network plot of most probable structures.
}
\description{
The function plots the network of the most probable structures.
}
\details{
This function uses \code{qgraph} for the visualization of the network. 
Additional arguments from the \code{qgraph} package can be passed on within this function.
}
