% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_structure_probability}
\alias{plot_structure_probability}
\title{Plot posterior likelihood distribution of all structures}
\usage{
plot_structure_probability(output, as.BF = TRUE)
}
\arguments{
\item{output}{Output of the rbinnet \code{select_structure} function}

\item{as.BF}{Determines if the raw posterior probabilities are plotted or the transformed Bayes Factor}
}
\value{
Plot of the posterior likelihood of the structures.
}
\description{
The function is based in the ggplot2 environment and plots the posterior likelihood
of the structures sorted from the most to the least probable. The function can either
plot the raw posterior probabilities or transformed as Bayes Factors depicting the 
Bayes Factor of the respective structure against the most probable structure.
}
\details{
Additional arguments can be passed to the function using by adding + and the common 
ggplot2 commands.
}
