% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emvs_functions.R
\name{screen_edges}
\alias{screen_edges}
\title{Bayesian edge screening for the Ising model.}
\usage{
screen_edges(
  x = x,
  theta = 0.5,
  precision = 0.975,
  prior_var_intercepts = 1,
  hierarchical = FALSE,
  alpha = 1,
  beta = 1,
  iteration_max = 100
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{theta}{The prior inclusion probability. The value \code{theta = 0.5},
in combination with \code{hierarchical = FALSE} stipulates a uniform prior
on the space of network structures.}

\item{precision}{A number between zero and one. The prior precision that is
desired for edge selection. Equal to one minus the desired type-1 error.
Defaults to \code{.975}.}

\item{prior_var_intercepts}{The variance of the prior distribution
stipulated on the Ising model's threshold parameters. Currently a normal
distribution is used for all threshold parameters, with a mean equal to
zero and a variance equal to \code{prior_var_intercepts}. Defaults to
\code{1}.}

\item{hierarchical}{Logical. If TRUE, a beta prior distribution is
imposed on the prior inclusion probability \code{theta} with
hyperparameters \code{alpha} and \code{beta}. A uniform prior on the
inclusion probability, a beta with \code{alpha = beta = 1}, stipulates a
uniform prior on network structure complexity.}

\item{alpha, beta}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}

\item{iteration_max}{The maximum number of Newton-Raphson iterations used in
optimization. Defaults to \code{1e2}. A warning is issued if procedure has
not converged in \code{iteration_max} iterations.}
}
\value{
A list containing the \code{p} by \code{p} matrices \code{sigma} and
  \code{gamma}. The matrix \code{sigma} is a numeric matrix with pairwise
  association estimates on the off-diagonal elements and threshold
  estimates on the diagonal elements. The matrix \code{gamma} contains the
  expected values of edge inclusion variables (i.e., the local posterior
  probability of edge inclusion. If \code{hierarchical = TRUE}, the modal
  estimate of the prior inclusion probability \code{theta} is also provided.
}
\description{
The function \code{screen_edges} screens promising edges for the Ising
model, using the joint pseudolikelihood and a spike and slab prior
distribution stipulated on the Ising model's association parameters.
}
\details{
The function \code{\link{screen_edges}} is a wrapper for
\code{edge_screening_emvs}.
}
\examples{
library("IsingSampler")
### Simulate dataset ###
# Input:
p <- 6 # Number of nodes
n <- 1000 # Number of samples
# Ising parameters:
Graph <- matrix(data = 0, nrow = p, ncol = p)
Graph[lower.tri(Graph)] <- rbinom(n = p * (p - 1) / 2,
                                  size = 1,
                                  prob = 0.2)
Graph <- Graph * runif(n = p ^ 2, min = 0.5, max = 2)
Graph <- Graph + t(Graph)
Thresholds <- -rowSums(Graph) / 2
# Simulate:
Data <- IsingSampler(n = n, graph = Graph, thresholds = Thresholds)
### Fit using fit_pseudoposterior ###
screened <- screen_edges(x = Data)
graph <- screened$sigma
graph[screened$gamma < 0.5] <- 0
# Plot results:
library("qgraph")
layout(t(1:2))
qgraph(graph,fade = FALSE)
title("Network of screened edges")
qgraph(Graph,fade = FALSE)
title("Original network")
}
