% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvs_functions.R
\name{select_structure}
\alias{select_structure}
\title{Bayesian structure selection for the Ising model.}
\usage{
select_structure(
  x,
  spike_var,
  slab_var,
  theta = 0.5,
  alpha = 1,
  beta = 1,
  hierarchical = FALSE,
  number_iterations = 1e+05,
  number_burnin_iterations = 100,
  prior_var_intercepts = 1,
  output_samples = FALSE,
  sigma,
  include,
  precision = 0.975,
  components = "normal",
  df = 5
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{spike_var, slab_var}{The \code{p} by \code{p} matrices of variances
that are used in the specification of the spike and slab prior
distributions that are stipulated on the association parameters. Optional,
if \code{spike_var} and \code{slab_var} are not specified, it makes use of
\code{precision} to specify \code{spike_var} and \code{slab_var} using
\code{\link{set_spike_and_slab}}.}

\item{theta}{The prior inclusion probability. The value \code{theta = 0.5},
in combination with \code{hierarchical = FALSE} stipulates a uniform prior
on the space of network structures. Defaults to \code{0.5}.}

\item{alpha}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}

\item{beta}{The hyperparameters of the beta prior distribution
stipulated on the prior inclusion probability \code{theta} if
\code{hierarchical = TRUE}. Default to \code{1}.}

\item{hierarchical}{Logical. If TRUE, a beta prior distribution is
imposed on the prior inclusion probability \code{theta} with
hyperparameters \code{alpha} and \code{beta}. A uniform prior on the
inclusion probability, a beta with \code{alpha = beta = 1}, stipulates a
uniform prior on network structure complexity.}

\item{number_iterations}{Integer. The number of iterations the Gibbs sampler
uses to explore the structure space. Defaults to \code{1e5}.}

\item{number_burnin_iterations}{Integer. The number of burnin iterations for
the Gibbs sampler. The output for the \code{number_burnin_iterations} is
not recorded. After \code{number_burnin_iterations}, the Gibbs sampler
runs for \code{number_iterations} iterations. Defaults to \code{0}.}

\item{prior_var_intercepts}{The variance of the prior distribution
stipulated on the Ising model's threshold parameters. Currently a normal
distribution is used for all threshold parameters, with a mean equal to
zero and a variance equal to \code{prior_var_intercepts}. Defaults to
\code{1}.}

\item{output_samples}{Logical. If \code{output_samples = TRUE},
\code{structure_selection_ssvs} returns all posterior samples of the Ising
model parameters. If \code{output_samples = FALSE},
\code{structure_selection_ssvs} returns the posterior means instead.}

\item{sigma}{A \code{p} by \code{p} numeric matrix with pairwise association
estimates on the off-diagonal elements and threshold estimates on the
diagonal elements. Optional. Can be used to specify starting values.}

\item{include}{A \code{p} by \code{p} binary matrix. If
\code{include[i,j] = 1}, the corresponding inclusion variable
\code{gamma[i, j]} is sampled. Otherwise, if \code{include[i,j] = 0}, the
inclusion variable \code{gamma[i,j]} and the association \code{sigma[i,j]}
are fixed to zero, which means that the edge between nodes \code{i}
and \code{j} is excluded from the structures that are considered.
Used to screen promising edges with \code{\link{screen_edges}}.}

\item{precision}{A number between zero and one. The prior precision that is
desired for edge selection. Equal to one minus the desired type-1 error.
Needs to be specified if \code{spike_var} and \code{slab_var} are
unspecified. Defaults to \code{.975}.}

\item{components}{One of "normal", "t", or "laplace". The type of
distribution that is used in the spike and slab set-up. The variance
of the slab component is matched to \code{slab_var}. Defaults to
\code{components = "normal"}.}

\item{df}{Used with \code{components = "t"}. The prior degrees of
freedom used in the T distribution. Defaults to \code{df = 5}.}
}
\value{
A list that contains \code{structures}, an \code{S} by
  \code{p(p-1)/2} matrix of binary variables that encodes the \code{S}
  structures that were visited by the Gibbs sampler. Includes edge names
  for easy reference. Also included is \code{posterior_probability},
  the \code{S} by \code{1} matrix of posterior probabilities that matches
  the rows of \code{structures}. If \code{output_samples = TRUE} the list
  contains an \code{number_iterations} by \code{p(p + 1) / 2} matrix
  \code{sigma_samples} that contains samples from the posterior distribution
  of the Ising model parameters, and if \code{hierarchical = TRUE}, it also
  includes an \code{number_iterations} by \code{1} matrix of
  \code{theta_samples} which contains samples from the posterior
  distribution of the prior inclusion probability \code{theta}. Otherwise,
  if \code{output_samples = FALSE}, the list contains a \code{p(p + 1) / 2}
  by \code{1} matrix \code{sigma_eap} of posterior means of the Ising model
  parameters, and if \code{hierarchical = TRUE}, it also includes
  \code{theta_eap}, the posterior mean of the prior inclusion probability.
}
\description{
The function \code{select_structure} explores the space of possible
Ising network structures using the Gibbs sampler.
}
\details{
The function \code{structure_selection_ssvs} is used by
\code{select_structure} for exploring the space of network structures. It
makes use of a Bayesian spike and slab approach to edge inclusion, and
explores the posterior space of network structures using the Gibbs sampler.
}
\examples{
\dontrun{
  library("IsingSampler")
  ### Simulate dataset ###
  # Input:
  p <- 6 # Number of nodes
  n <- 1000 # Number of samples
  # Ising parameters:
  Graph <- matrix(data = 0, nrow = p, ncol = p)
  Graph[lower.tri(Graph)] <- rbinom(n = p * (p - 1) / 2,
                                    size = 1,
                                    prob = 0.2)
  Graph <- Graph * runif(n = p ^ 2, min = 0.5, max = 2)
  Graph <- Graph + t(Graph)
  Thresholds <- -rowSums(Graph) / 2
  # Simulate:
  Data <- IsingSampler(n = n, graph = Graph, thresholds = Thresholds)
  ### Fit using fit_pseudoposterior ###
  selection <- select_structure(x = Data,
                                number_iterations = 1e3,
                                number_burnin_iterations = 0,
                                hierarchical = FALSE)
  incl_prob <- matrix(0, nrow = p, ncol = p)
  incl_prob[lower.tri(incl_prob, diag = FALSE)] <-
    t(selection$structures) \%*\% selection$posterior_probability
  incl_prob <- incl_prob + t(incl_prob)
  mps <- matrix(0, nrow = p, ncol = p)
  mps[lower.tri(mps, diag = TRUE)] <- selection$sigma_eap
  mps <- mps + t(mps)
  mps[incl_prob < 0.5] <- 0
  # Plot results:
  library("qgraph")
  layout(t(1:2))
  qgraph(mps,fade = FALSE)
  title("Median probability structure")
  qgraph(Graph,fade = FALSE)
  title("Original network")
}
}
