% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{set_spike_and_slab}
\alias{set_spike_and_slab}
\title{Specification of the spike and slab variance parameters.}
\usage{
set_spike_and_slab(x, precision)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix containing binary coded
responses (i.e., coded \code{0,1}) for \code{n} independent observations
on \code{p} variables in the network or graph.}

\item{precision}{A number between zero and one. The prior precision that is
desired for edge selection. Equal to one minus the desired type-1 error.}
}
\value{
A list containing \code{spike_var} and \code{slab_var}, which are
  \code{p} by \code{p} matrices of variances used for the spike and slab
  prior components, \code{sigma_ml}, a \code{p} by \code{p} matrix of
  maximum pseudolikelihood estimates, with association estimates on the off-
  diagonal and threshold estimates on the diagonal.
}
\description{
The function \code{set_spike_and_slab} is used to specify the variance
parameters of the spike and slab prior distribution that is stipulated
on the association parameters of the Ising model. It is used by
\code{\link{edge_screening}} and \code{\link{structure_selection}}.
}
\details{
The function \code{set_spike_and_slab} runs
\code{\link{optimize_pseudoposterior}} to produce \code{slab_var},
a \code{p} by \code{p} matrix of unit information prior variances. It
furthermore aims to find the root of \code{\link{xi_delta_matching}} to
match the spike and slab intersections to a predefined precision. If
succesfull, \code{spike_var} contains the corresponding variances of the
spike components.
}
